Split JustFugitives into multiple files
