+++ app/src/layout/JustFugitives/FugitiveListItem.css
@@ -1,0 +1,92 @@
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item {
+    height: 20rem;
+    width: 20rem;
+    background-size: cover;
+}
+
+/* This is setup such that the fugitive popover is present, but only
+   displays the fugitive's name at the bottom of the image */
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item .fugitive-popover {
+    -webkit-backdrop-filter: blur(1rem);
+    backdrop-filter: blur(1rem);
+    width: 100%;
+    padding: 0.5rem;
+    box-sizing: border-box;
+    transition: 0.25s all;
+    height: 2.5rem;
+    overflow: hidden;
+}
+
+.dark .fbi-most-wanted-page .fugitives-list .fugitive-list-item .fugitive-popover {
+    background: rgba(0, 0, 0, 0.5);
+}
+
+.light .fbi-most-wanted-page .fugitives-list .fugitive-list-item .fugitive-popover {
+    background: rgba(255, 255, 255, 0.5);
+}
+
+/* When the popover is hovered, all content becomes centered */
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item .fugitive-popover:hover {
+    justify-content: center !important;
+}
+
+/* Expands the popover to the full height of the image */
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item:hover .fugitive-popover {
+    height: 100%;
+}
+
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item .fugitive-popover h3.fugitive-name {
+    width: 100%;
+    padding: 0 1rem;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    font-weight: 200;
+    margin: 0;
+}
+
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item:hover .fugitive-popover h3.fugitive-name {
+    font-weight: 600;
+    white-space: normal;
+    overflow: visible;
+    text-align: center;
+}
+
+.fbi-most-wanted-page .fugitives-list .fugitive-list-item .fugitive-popover .fugitive-poster-link {
+    display: none;
+    border: none;
+    font-size: 0.75rem;
+    border-radius: 0.5rem;
+    text-decoration: none;
+    padding: 0.5rem;
+    box-sizing: border-box;
+    margin-top: 0.5rem;
+}
+
+.dark
+    .fbi-most-wanted-page
+    .fugitives-list
+    .fugitive-list-item
+    .fugitive-popover
+    .fugitive-poster-link {
+    color: var(--primary-text-dark);
+    background-color: rgba(0, 0, 0, 0.5);
+}
+
+.light
+    .fbi-most-wanted-page
+    .fugitives-list
+    .fugitive-list-item
+    .fugitive-popover
+    .fugitive-poster-link {
+    color: var(--primary-text-light);
+    background-color: rgba(255, 255, 255, 0.5);
+}
+
+.fbi-most-wanted-page
+    .fugitives-list
+    .fugitive-list-item:hover
+    .fugitive-popover
+    .fugitive-poster-link {
+    display: block;
+}
